Make better use of CLOS
There were a bunch of methods before, like FORMAT-EMAIL, that are more
elegantly expressed using specialized types and methods. By creating a
new class for any type that needs to be formatted differently, there
is no need for individual formatters to have knowledge of the types
they're formatting; it is only necessary to convert the resume before
formatting into a structured form where the data is in a typed form.